API added
Added an API request in the "history" apps tab that gets the currency values in a certain date with a certain base currency. next i have to "restart" the app, maybe by creating a button inside the results display.  it should clear the results display, hide it, clear the form and get the user back to the form,